# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anvilml-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.27"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List models",
        "description": "Lists all discovered models, optionally filtered by architecture kind.",
        "operationId": "list_models",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "Filter results to only models of this kind.\n\nWhen `None`, all models are returned regardless of kind.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ModelKind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelMeta"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
...
    "/v1/models/{id}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get a model",
        "description": "Looks up a single model by its ID.",
        "operationId": "get_model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMeta"
                }
              }
            }
          },
          "404": {
            "description": "Model not found"
          }
        }
      }
    },
...
    "/v1/workers/{id}/restart": {
      "post": {
        "tags": [
          "Workers"
        ],
        "summary": "Restart a worker",
        "description": "Restarts the worker identified by ID: graceful shutdown followed by respawn into the same device slot.",
        "operationId": "restart_worker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Worker ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Restart accepted"
          },
          "404": {
            "description": "Worker not found"
          },
          "409": {
            "description": "Worker already shutting down"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ModelDtype": {
        "type": "string",
        "description": "The data type or precision of a model's weights.\n\nThis enum covers the most common weight precisions encountered in\ngenerative-AI model files, from full FP32 through FP4 quantisation.",
        "enum": [
          "fp32",
          "fp16",
          "bf16",
          "fp8",
          "fp4",
          "unknown"
        ]
      },
      "ModelFormat": {
        "type": "string",
        "description": "The storage format of a model file.\n\nEach variant names a file format used to serialise model weights\non disk. The `Unknown` variant is used when the file extension or\nheader does not match any recognised format.",
        "enum": [
          "safetensors",
          "ckpt",
          "pt",
          "bin",
          "unknown"
        ]
      },
      "ModelKind": {
        "type": "string",
        "description": "The architecture family of a model file.\n\nEach variant names a distinct role a model plays in the AnvilML\ncomputation graph — diffusion base model, text encoder, VAE, LoRA\nadapter, ControlNet, or upscaler.",
        "enum": [
          "diffusion",
          "text_encoder",
          "vae",
          "lora",
          "control_net",
          "upscale",
          "unknown"
        ]
      },
      "ModelMeta": {
        "type": "object",
        "description": "Metadata about a discovered model file.\n\nThis struct is the primary output of the model scanner: it captures\nthe stable identity, location, architecture family, data type, file\nformat, size, modification time, and scan timestamp of a single\nmodel file on disk.",
        "required": [
          "id",
          "name",
          "path",
          "kind",
          "dtype",
          "format",
          "size_bytes",
          "mtime_unix",
          "scanned_at"
        ],
        "properties": {
          "dtype": {
            "$ref": "#/components/schemas/ModelDtype",
            "description": "The model's data type / precision."
          },
          "format": {
            "$ref": "#/components/schemas/ModelFormat",
            "description": "The model file format."
          },
          "id": {
            "type": "string",
            "description": "Stable identifier: SHA256 hex of the first 1 MiB of the file."
          },
          "kind": {
            "$ref": "#/components/schemas/ModelKind",
            "description": "The model's architecture family."
          },
          "mtime_unix": {
            "type": "integer",
            "format": "int64",
            "description": "File modification time as Unix epoch seconds (populated by the scanner)."
          },
          "name": {
            "type": "string",
            "description": "Human-readable model name."
          },
          "path": {
            "type": "string",
            "description": "Filesystem path to the model file."
          },
          "scanned_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when this metadata was scanned."
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "File size in bytes.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}